ppt/my-ai-ppt/apps/api/requirements.txt
fastapi
uvicorn
pydantic
python-pptx
